updated for TF 1.0
https://www.tensorflow.org/install/migration#
