+++ models/dsc.py
@@ -4,8 +4,6 @@ import tensorflow as tf
 import utils.plot_functions as pf
 from models.base_model import Model
 
-import os
-
 class dsc(Model):
   def __init__(self, params, schedule):
     Model.__init__(self, params, schedule)
@@ -15,6 +13,7 @@ class dsc(Model):
   Inputs:
    params: [dict] model parameters
   Modifiable Parameters:
+    rectify_u      [bool] If set, rectify layer 1 activity
     rectify_v      [bool] If set, rectify layer 2 activity
     norm_weights   [bool] If set, l2 normalize weights after updates
     batch_size     [int] Number of images in a training batch
@@ -28,6 +27,7 @@ class dsc(Model):
     # Meta parameters
     self.rectify_u = bool(params["rectify_u"])
     self.rectify_v = bool(params["rectify_v"])
+    self.norm_weights = bool(params["norm_weights"])
     # Network Size
     self.batch_size = int(params["batch_size"])
     self.num_pixels = int(params["num_pixels"])
@@ -35,7 +35,6 @@ class dsc(Model):
     self.num_v = int(params["num_v"])
     self.a_shape = [self.num_pixels, self.num_u]
     self.b_shape = [self.num_u, self.num_v]
-    self.w_shapes = [vals for vals in zip(self.a_shape, self.b_shape)]
     # Hyper Parameters
     self.num_steps = int(params["num_steps"])
 
@@ -43,7 +42,7 @@ class dsc(Model):
   def check_params(self):
     Model.check_params(self)
     assert np.sqrt(self.num_u) == np.floor(np.sqrt(self.num_u)), (
-      "The parameter `num_u` must have an even square-root for plotting.")
+      "The parameter `num_u` must have an even square-root.")
 
   """
   Returns total loss function for given input
@@ -53,54 +52,36 @@ class dsc(Model):
     input_data []
     u_state []
     v_state []
+    a []
+    b []
   """
-  def compute_loss(self, input_data, u_state, v_state):
+  def compute_loss(self, input_data, u_state=None, v_state=None, a=None,
+    b=None):
+    with tf.variable_scope("layers", reuse=True) as scope:
+      if u_state is None:
+        u_state = tf.get_variable("u")
+      if v_state is None:
+        v_state = tf.get_variable("v")
     with tf.variable_scope("weights", reuse=True) as scope:
-      a_state = tf.get_variable(name="a")
-      b_state = tf.get_variable(name="b")
-    with tf.name_scope("iterative_loss"):
-      temp_sigma = tf.exp(tf.matmul(v_state, tf.transpose(b_state)))#0.83
-      temp_x_ = tf.matmul(u_state, tf.transpose(a_state))
-      self.input_stddev = 1.0#tf.reduce_mean(tf.square(tf.nn.moments(input_data,
-        #axes=[1])[1]))
-      recon_loss = tf.multiply(self.recon_mult, tf.reduce_mean(
-        tf.multiply(tf.divide(1.0, (2.0*self.input_stddev)),
-        tf.reduce_sum(tf.square(tf.subtract(input_data, temp_x_)),
-        axis=[1]))), name="recon_loss")
-      feedback_loss = tf.reduce_mean(tf.reduce_sum(tf.add(
-        tf.divide(u_state, temp_sigma), tf.log(temp_sigma)),
-        axis=[1]), name="fb_loss")
-      sparse_loss = tf.multiply(self.sparse_mult, tf.reduce_mean(
-        tf.reduce_sum(tf.abs(v_state), axis=[1])), name="sparse_loss")
-      a_loss = tf.multiply(self.a_decay_mult, tf.reduce_mean(
-        tf.reduce_sum(tf.square(a_state), axis=[1])), name="a_loss")
-      b_loss = tf.multiply(self.b_decay_mult, tf.reduce_mean(
-        tf.reduce_sum(tf.abs(b_state), axis=[1])), name="b_loss")
-      total_loss = tf.add_n([recon_loss, feedback_loss, sparse_loss,
-        a_loss, b_loss], name="total_loss")
-    return (total_loss, recon_loss, feedback_loss, sparse_loss, a_loss, b_loss)
-
-  def inference_step(self, inference_idx, u_in, v_in, u_relu, v_relu):
-    current_u_loss = self.compute_loss(self.x, u_in, v_in)[0]
-    with tf.name_scope("inference"+str(inference_idx)):
-      u_grad = tf.gradients(current_u_loss, u_in)[0]
-      u_update = tf.add(u_in, tf.multiply(-self.u_step_size, u_grad),
-        name="u_update"+str(inference_idx))
-      u_out = tf.nn.relu(u_update) if u_relu else u_update
-    current_v_loss = self.compute_loss(self.x, u_out, v_in)[0]
-    with tf.name_scope("inference"+str(inference_idx)):
-      v_grad = tf.gradients(current_v_loss, v_in)[0]
-      v_update = tf.add(v_in, tf.multiply(-self.v_step_size, v_grad),
-        name="v_update"+str(inference_idx))
-      v_out = tf.nn.relu(v_update) if v_relu else v_update
-      return u_out, v_out
+      if a is None:
+        a = tf.get_variable(name="a")
+      if b is None:
+        b = tf.get_variable(name="b")
+    temp_sigma = tf.exp(tf.matmul(v_state, tf.transpose(b)))
+    temp_x_ = tf.matmul(u_state, tf.transpose(a))
+    recon_loss = self.recon_mult * tf.reduce_mean(0.5 *
+      tf.reduce_sum(tf.pow(tf.subtract(input_data, temp_x_), 2.0), axis=[1]))
+    feedback_loss = tf.reduce_mean(tf.reduce_sum(tf.add(
+      tf.divide(u_state, temp_sigma), tf.log(temp_sigma)), axis=[1]))
+    sparse_loss = self.sparse_mult * tf.reduce_mean(
+      tf.reduce_sum(tf.abs(v_state), axis=[1]))
+    total_loss = (recon_loss + feedback_loss + sparse_loss)
+    return (total_loss, recon_loss, feedback_loss, sparse_loss)
 
   """
-  Based on the neworks described in:
+  Build the nework described in:
     Y Karklin, MS Lewicki (2005) - A Hierarchical Bayesian Model for Learning
       Nonlinear Statistical Regularities in Nonstationary Natural Signals
-    TS Lee, D Mumford (2003) - Hierarchical Bayesian Inference in the Visual
-      Cortex
   Method for unrolling inference into the graph
     # compute loss for current sate
     # take gradient of loss wrt current state
@@ -117,71 +98,88 @@ class dsc(Model):
             shape=(), name="recon_mult")
           self.sparse_mult = tf.placeholder(tf.float32,
             shape=(), name="sparse_mult")
-          self.a_decay_mult = tf.placeholder(tf.float32,
-            shape=(), name="a_decay_mult")
-          self.b_decay_mult = tf.placeholder(tf.float32,
-            shape=(), name="b_decay_mult")
           self.u_step_size = tf.placeholder(tf.float32,
             shape=(), name="u_step_size")
           self.v_step_size = tf.placeholder(tf.float32,
             shape=(), name="v_step_size")
 
         with tf.name_scope("constants") as scope:
-          self.u_noise = tf.truncated_normal(
+          self.u_zeros = tf.zeros(
             shape=tf.stack([tf.shape(self.x)[0], self.num_u]),
-            mean=0.0, stddev=0.1, dtype=tf.float32, name="u_noise")
-          self.v_ones = tf.ones(
+            dtype=tf.float32, name="u_zeros")
+          self.v_zeros = tf.zeros(
             shape=tf.stack([tf.shape(self.x)[0], self.num_v]),
-            dtype=tf.float32, name="v_ones")
+            dtype=tf.float32, name="v_zeros")
 
         with tf.name_scope("step_counter") as scope:
           self.global_step = tf.Variable(0, trainable=False,
             name="global_step")
 
-        w_inits = [np.load(os.path.expanduser("~")+"/Work/Projects/pretrain/analysis/0.0/weights/phi.npz")["data"],
-          tf.multiply(0.1, tf.ones(self.w_shapes[1], dtype=tf.float32),
-          name="b_init")]
-
         with tf.variable_scope("weights") as scope:
           self.a = tf.get_variable(name="a", dtype=tf.float32,
-            initializer=w_inits[0], trainable=True)
+            initializer=tf.truncated_normal(self.a_shape, mean=0.0,
+            stddev=1.0, dtype=tf.float32, name="a_init"), trainable=True)
           self.b = tf.get_variable(name="b", dtype=tf.float32,
-            initializer=w_inits[1], trainable=True)
+            initializer=tf.truncated_normal(self.b_shape, mean=0.0,
+            stddev=1.0, dtype=tf.float32, name="b_init"), trainable=True)
 
-        u_list = [self.u_noise]
-        v_list = [self.v_ones]
-        for inference_idx in range(self.num_steps-1):
-          u_out, v_out = self.inference_step(inference_idx,
-            u_list[inference_idx], v_list[inference_idx],
-            self.rectify_u, self.rectify_v)
-          u_list.append(u_out)
-          v_list.append(v_out)
+        with tf.name_scope("norm_weights") as scope:
+          self.norm_a = self.a.assign(tf.nn.l2_normalize(self.a,
+            dim=0, epsilon=self.eps, name="row_l2_norm"))
+          self.norm_b = self.b.assign(tf.nn.l2_normalize(self.b,
+            dim=0, epsilon=self.eps, name="row_l2_norm"))
+          self.norm_weights = tf.group(self.norm_a, self.norm_b,
+            name="l2_normalization")
 
         with tf.variable_scope("layers") as scope:
           self.u = tf.get_variable(name="u", dtype=tf.float32,
-            initializer=u_list[-1], trainable=False, validate_shape=False)
+            initializer=self.u_zeros, trainable=False, validate_shape=False)
           self.v = tf.get_variable(name="v", dtype=tf.float32,
-            initializer=v_list[-1], trainable=False, validate_shape=False)
+            initializer=self.v_zeros, trainable=False, validate_shape=False)
+
         with tf.name_scope("output") as scope:
           with tf.name_scope("image_estimate"):
-            self.x_ = tf.matmul(self.u, tf.transpose(self.a),
-              name="reconstruction")
+            self.x_ = tf.matmul(self.u, tf.transpose(self.a), name="reconstruction")
             MSE = tf.reduce_mean(tf.pow(tf.subtract(self.x, self.x_), 2.0),
               axis=[1, 0], name="mean_squared_error")
-            self.pSNRdB = tf.multiply(10.0, tf.log(tf.divide(tf.square(1.0),
-              MSE)), name="recon_quality")
+            self.pSNRdB = tf.multiply(10.0, tf.log(tf.divide(tf.pow(1.0,
+               2.0), MSE)), name="recon_quality")
           with tf.name_scope("layer1_prior"):
-            self.sigma = tf.exp(tf.matmul(self.v, tf.transpose(self.b),
+            self.sigma = tf.exp(-tf.matmul(self.v, tf.transpose(self.b),
               name="sigma"))
 
-          with tf.name_scope("loss") as scope:
-            with tf.name_scope("unsupervised"):
-              (self.total_loss,
-              self.recon_loss,
-              self.feedback_loss,
-              self.sparse_loss,
-              self.a_loss,
-              self.b_loss) = self.compute_loss(self.x, self.u, self.v)
+        with tf.name_scope("loss") as scope:
+          with tf.name_scope("unsupervised"):
+            (self.total_loss,
+            self.recon_loss,
+            self.feedback_loss,
+            self.sparse_loss) = self.compute_loss(self.x)
+
+        with tf.name_scope("inference") as scope:
+          self.clear_u = self.u.assign(self.u_zeros)
+          self.clear_v = self.v.assign(self.v_zeros)
+          self.reset_activity = tf.group(self.clear_u, self.clear_v,
+            name="reset_activity")
+          current_loss = []
+          self.u_t = [self.u_zeros] # init to zeros
+          self.v_t = [self.v_zeros] # init to zeros
+          for step in range(self.num_steps-1): # loop doesn't include init
+            current_loss.append(self.compute_loss(self.x, self.u_t[step],
+              self.v_t[step])[0])
+            du = -tf.gradients(current_loss[step], self.u_t[step])[0]
+            if self.rectify_u:
+              new_u = tf.nn.relu(self.u_t[step] + self.u_step_size * du)
+            else:
+              new_u = self.u_t[step] + self.u_step_size * du
+            self.u_t.append(new_u)
+            dv = -tf.gradients(current_loss[step], self.v_t[step])[0]
+            if self.rectify_v:
+              new_v = tf.nn.relu(self.v_t[step] + self.v_step_size * dv)
+            else:
+              new_v = self.v_t[step] + self.v_step_size * dv
+            self.v_t.append(new_v)
+          self.full_inference = tf.group(self.u.assign(self.u_t[-1]),
+            self.v.assign(self.v_t[-1]), name="full_inference")
 
     self.graph_built = True
 
@@ -204,8 +202,6 @@ class dsc(Model):
     recon_loss = np.array(self.recon_loss.eval(feed_dict)).tolist()
     feedback_loss = np.array(self.feedback_loss.eval(feed_dict)).tolist()
     sparse_loss = np.array(self.sparse_loss.eval(feed_dict)).tolist()
-    a_loss = np.array(self.a_loss.eval(feed_dict)).tolist()
-    b_loss = np.array(self.b_loss.eval(feed_dict)).tolist()
     total_loss = np.array(self.total_loss.eval(feed_dict)).tolist()
     u_vals = tf.get_default_session().run(self.u, feed_dict)
     u_vals_max = np.array(u_vals.max()).tolist()
@@ -222,8 +218,6 @@ class dsc(Model):
       "recon_loss":recon_loss,
       "feedback_loss":feedback_loss,
       "sparse_loss":sparse_loss,
-      "a_l2_loss":a_loss,
-      "b_l1_loss":b_loss,
       "total_loss":total_loss,
       "u_max":u_vals_max,
       "v_max":v_vals_max,
@@ -236,7 +230,6 @@ class dsc(Model):
       stat_dict[name+"_min_grad"] = np.array(grad.min()).tolist()
     js_str = js.dumps(stat_dict, sort_keys=True, indent=2)
     self.log_info("<stats>"+js_str+"</stats>")
-    #print(self.input_stddev.eval(feed_dict))
 
   """
   Plot weights, reconstruction, and gradients
@@ -248,50 +241,18 @@ class dsc(Model):
     Model.generate_plots(self, input_data, input_labels)
     feed_dict = self.get_feed_dict(input_data, input_labels)
     current_step = str(self.global_step.eval())
-    recon = tf.get_default_session().run(self.x_, feed_dict)
-    a_weights = tf.get_default_session().run(self.a, feed_dict)
-    #b_weights = tf.get_default_session().run(self.b, feed_dict)
-    u_vals = tf.get_default_session().run(self.u, feed_dict)
-    #v_vals = tf.get_default_session().run(self.v, feed_dict)
-    #pf.save_data_tiled(input_data.reshape((self.batch_size,
-    #  np.int(np.sqrt(self.num_pixels)),
-    #  np.int(np.sqrt(self.num_pixels)))),
-    #  normalize=False, title="Images at step "+current_step,
-    #  save_filename=(self.disp_dir+"images_"+self.version+"-"
-    #  +current_step.zfill(5)+".pdf"), vmin=np.min(input_data),
-    #  vmax=np.max(input_data))
-    pf.save_data_tiled(recon.reshape((self.batch_size,
-      np.int(np.sqrt(self.num_pixels)),
-      np.int(np.sqrt(self.num_pixels)))),
-      normalize=False, title="Recons at step "+current_step,
-      save_filename=(self.disp_dir+"recons_v"+self.version+"-"
-      +current_step.zfill(5)+".pdf"), vmin=np.min(recon), vmax=np.max(recon))
-    pf.save_data_tiled(a_weights.T.reshape(self.num_u,
+    pf.save_data_tiled(
+      tf.transpose(self.b).eval().T.reshape(self.num_v,
+      int(np.sqrt(self.num_u)), int(np.sqrt(self.num_u))),
+      normalize=True, title="Density weights matrix at step number "
+      +current_step, save_filename=(self.disp_dir+"b_v"+self.version+"-"
+      +current_step.zfill(5)+".pdf"))
+    pf.save_data_tiled(
+      tf.transpose(self.a).eval().reshape(self.num_u,
       int(np.sqrt(self.num_pixels)), int(np.sqrt(self.num_pixels))),
-      normalize=False, title="Dictionary at step "+current_step,
+      normalize=True, title="Dictionary at step "+current_step,
       save_filename=(self.disp_dir+"a_v"+self.version+"-"
       +current_step.zfill(5)+".pdf"))
-    #pf.save_data_tiled(b_weights.T.reshape(self.num_v,
-    #  int(np.sqrt(self.num_u)), int(np.sqrt(self.num_u))),
-    #  normalize=False, title="Density weights matrix at step number "
-    #  +current_step, save_filename=(self.disp_dir+"b_v"+self.version+"-"
-    #  +current_step.zfill(5)+".pdf"))
-    pf.save_activity_hist(u_vals, num_bins=1000,
-      title="u Activity Histogram at step "+current_step,
-      save_filename=(self.disp_dir+"u_hist_v"+self.version+"-"
-      +current_step.zfill(5)+".pdf"))
-    #pf.save_activity_hist(v_vals, num_bins=1000,
-    #  title="v Activity Histogram at step "+current_step,
-    #  save_filename=(self.disp_dir+"v_hist_v"+self.version+"-"
-    #  +current_step.zfill(5)+".pdf"))
-    pf.save_bar(np.linalg.norm(a_weights, axis=1, keepdims=False), num_xticks=5,
-      title="a l2 norm", save_filename=(self.disp_dir+"a_norm_v"+self.version
-      +"-"+current_step.zfill(5)+".pdf"), xlabel="Basis Index",
-      ylabel="L2 Norm")
-    #pf.save_bar(np.linalg.norm(b_weights, axis=1, keepdims=False), num_xticks=5,
-    #  title="b l2 norm", save_filename=(self.disp_dir+"b_norm_v"+self.version
-    #  +"-"+current_step.zfill(5)+".pdf"), xlabel="Basis Index",
-    #  ylabel="L2 Norm")
     for weight_grad_var in self.grads_and_vars[self.sched_idx]:
       grad = weight_grad_var[0][0].eval(feed_dict)
       shape = grad.shape
@@ -299,12 +260,12 @@ class dsc(Model):
       if name == "a":
         pf.save_data_tiled(grad.T.reshape(self.num_u,
           int(np.sqrt(self.num_pixels)), int(np.sqrt(self.num_pixels))),
-          normalize=False, title="Gradient for a at step "+current_step,
+          normalize=True, title="Gradient for a at step "+current_step,
           save_filename=(self.disp_dir+"da_v"+self.version+"_"
           +current_step.zfill(5)+".pdf"))
-      #elif name == "b":
-      #  pf.save_data_tiled(grad.T.reshape(self.num_v,
-      #    int(np.sqrt(self.num_u)), int(np.sqrt(self.num_u))),
-      #    normalize=False, title="Gradient for b at step "+current_step,
-      #    save_filename=(self.disp_dir+"db_v"+self.version+"_"
-      #    +current_step.zfill(5)+".pdf"))
+      elif name == "b":
+        pf.save_data_tiled(grad.T.reshape(self.num_v,
+          int(np.sqrt(self.num_u)), int(np.sqrt(self.num_u))),
+          normalize=True, title="Gradient for b at step "+current_step,
+          save_filename=(self.disp_dir+"db_v"+self.version+"_"
+          +current_step.zfill(5)+".pdf"))
